Fix chart

+++ src/utils/storageUtil.js
@@ -1,10 +1,12 @@
 /*global chrome*/
+const isChromeLocalStorage = chrome && chrome.local;
+
 export function Get(key, callback) {
-    if (typeof key === "undefined") {
+    if (typeof key === 'undefined') {
         return;
     }
     const data = window.localStorage.getItem(key);
-    if (callback) {
+    if (callback && isChromeLocalStorage) {
         chrome.storage.local.get(key, callback);
     }
     try {
@@ -15,27 +17,36 @@ export function Get(key, callback) {
 }
 
 export function Set(key, value, callback) {
-    if (typeof key === "undefined" || typeof value === "undefined") {
+    if (typeof key === 'undefined' || typeof value === 'undefined') {
         return;
     }
     if (typeof value !== 'string') {
         value = JSON.stringify(value);
     }
     window.localStorage.setItem(key, value);
-    chrome.storage.local.set({
-        [key]: value
-    }, callback);
+    if (isChromeLocalStorage) {
+        chrome.storage.local.set(
+            {
+                [key]: value,
+            },
+            callback
+        );
+    }
 }
 
 export function Remove(key, callback) {
-    if (typeof key === "undefined") {
+    if (typeof key === 'undefined') {
         return;
     }
     window.localStorage.removeItem(key);
-    chrome.storage.local.remove([key], callback);
+    if (isChromeLocalStorage) {
+        chrome.storage.local.remove([key], callback);
+    }
 }
 
 export function Clear(callback) {
     window.localStorage.clear();
-    chrome.storage.local.clear(callback);
+    if (isChromeLocalStorage) {
+        chrome.storage.local.clear(callback);
+    }
 }
